requirements.txt
aiogram>=3.4.0
python-dotenv>=1.0.0
sqlalchemy>=2.0
aiosqlite>=0.19.0
openai>=1.0
